+++ examples/src/main/java/me/bechberger/minicli/examples/EnumsAndCompletionCandidates.java
@@ -1,8 +1,10 @@
 package me.bechberger.minicli.examples;
 
 import me.bechberger.minicli.MiniCli;
+import me.bechberger.minicli.annotations.Command;
 import me.bechberger.minicli.annotations.Option;
 
+@Command(name = "enums")
 public class EnumsAndCompletionCandidates implements Runnable {
     enum Mode { fast, safe }
 
@@ -12,6 +14,7 @@ public class EnumsAndCompletionCandidates implements Runnable {
     Mode mode;
 
     public void run() {
+        System.out.println("Mode: " + mode);
     }
 
     public static void main(String[] args) {