+++ examples/src/test/java/examples/PositionalParametersTest.java
@@ -1,0 +1,30 @@
+package examples;
+
+import me.bechberger.minicli.MiniCli;
+import me.bechberger.minicli.examples.PositionalParameters;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class PositionalParametersTest {
+    @Test
+    public void testHelp() {
+        var res = MiniCli.runCaptured(new PositionalParameters(), new String[]{"--help"});
+        assertEquals("""
+                Usage: positionalparameters [-hV] FILE [ARGS...]
+                      FILE         Input file
+                      [ARGS...]    Extra arguments
+                  -h, --help       Show this help message and exit.
+                  -V, --version    Print version information and exit.
+                """, res.out());
+    }
+
+    @Test
+    public void testPositionalParameters() {
+        var res = MiniCli.runCaptured(new PositionalParameters(), new String[]{"in.txt", "arg1", "arg2"});
+        assertEquals("""
+                File: in.txt
+                Args: [arg1, arg2]
+                """, res.out());
+    }
+}