+++ examples/src/test/java/examples/Util.java
@@ -1,17 +1,12 @@
-package me.bechberger.minicli.examples;
-
-import me.bechberger.minicli.RunResult;
-import org.junit.jupiter.api.Test;
+package examples;
 
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
 import java.lang.reflect.InvocationTargetException;
 
-import static org.junit.jupiter.api.Assertions.*;
+public class Util {
 
-public class ToolTest {
-
-    String run(Class<?> cls, String... args) {
+    static String run(Class<?> cls, String... args) {
         PrintStream originalOut = System.out;
         PrintStream originalErr = System.err;
         ByteArrayOutputStream outContent = new ByteArrayOutputStream();
@@ -26,22 +21,6 @@ public class ToolTest {
         System.setOut(originalOut);
         System.setErr(originalErr);
         return outContent + errContent.toString();
-    }
-
-    @Test
-    public void testCustomHeaderAndSynopsis() {
-        assertEquals("""
-                My Tool
-                Copyright 2026
-                Usage: mytool [OPTIONS] <file>
-                Process files
-                      --flag
-                  -h, --help       Show this help message and exit.
-                  -V, --version    Print version information and exit.
-                
-                Examples:
-                  mytool --flag
-                """, run(CustomHeaderAndSynopsis.class, "--help"));
     }
 
 }